requirements.txt
asgiref==3.2.3
astroid==2.3.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.3
Flask==1.1.1
future==0.18.2
gunicorn==19.3.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.2
line-bot-sdk==1.8.0
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.17.4
opencv-python==4.1.2.30
pandas==0.25.2
Pillow==7.0.0
psycopg2==2.8.4
pylint==2.4.3
pytesseract==0.3.1
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.2
qrcode==6.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
urllib3==1.25.8
users==1.0.dev0
virtualenv==16.7.9
Werkzeug==0.16.1
wrapt==1.11.2
